backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.24.1
pydantic==2.9.0
python-multipart==0.0.9
python-dotenv==1.0.0
supabase==2.3.0
sentry-sdk[fastapi]==1.40.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
ffmpeg-python==0.2.0
google-generativeai==0.8.3
